% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separableNNTCIs.R
\name{separableNNTCIs}
\alias{separableNNTCIs}
\title{Calculate 95\% CI for NNT based on separable effects}
\usage{
separableNNTCIs(dat, Y, eoiValue, crsValue, dTime, cutTimes, A, X, id, nboot)
}
\arguments{
\item{dat}{The data set used for the final analysis.}

\item{Y}{The outcome with 3 levels. By default, \code{Y=0} for censoring, \code{Y=1} for an event of interest, and \code{Y=2} for competing events.}

\item{eoiValue}{The value of \code{Y=1} for the event of interest by default.}

\item{crsValue}{The value of \code{Y=2} for the competing events by default.}

\item{dTime}{The time to events, including censoring, an event of interest, and competing events.}

\item{cutTimes}{The discrete time intervals for the pooled logistic regression models.}

\item{A}{The indicator for exposure or treatment with 2 levels. By default, \code{A=1} for the treatment group and \code{A=0} for the control group.}

\item{X}{The minimal sufficient of confounders considered in the final analysis,}

\item{id}{The unique ID for each observed included in \code{dat}.}

\item{nboot}{The number of bootstrapped samples. By default, \code{nboot=50}.}
}
\value{
A matrix list contains three data.frames
\describe{
  \item{datRessult}{The estimated NNT and relevant measurements}
  \item{datCumulativeIncidece}{The estimated cumulative incidence}
  \item{datBoots}{The bootstrapped distributions of NNT and relevant measurements}
}
}
\description{
To calculate the 95\% confidence intervals for Number Needed to Treat (NNT) based on the separable effects.
}
\references{
\enumerate{
  \item Yang et al. (2023) Disentangling the inverse LDL-C-hemorrhagic stroke association in Chinese adults with hypertension: Findings from the Chinese Multi-provincial Cohort Study. Am J Epidemiol. (Revision)
  \item Stensrud el al. (2022) Separable effects for causal inference in the presence of competing events. J Am Stat Assoc. 117(537):175-83. <doi: https://doi.org/10.1080/01621459.2020.1765783>
}
}
